Remove cleanup logic for older releases of Chef Infra Server
We required stepped upgrades so we know that everyone has first upgraded to 12.17. That means this logic all ran there and we can remove it here. This greatly reduces the amount of things we have to do each time the reconfigure, upgrade, or cleanup commands run and that speeds up runs / makes troubleshooting easier.

Signed-off-by: Tim Smith <tsmith@chef.io>